:sparkles: feat : 내 쿠폰 조회 기능 구현

+++ src/main/kotlin/kr/hhplus/be/server/api/coupon/response/CouponResponse.kt
@@ -9,7 +9,7 @@ data class CouponResponse(
     val name: String,
     val description: String,
     val discountType: CouponDiscountType,
-    val discountValue: Int,
+    val discountValue: Long,
     val status: CouponStatus,
 )
